feat(adaptor): Consider adaptor_connection (bool) when assigning rendermodel offsets
Pass the adaptor_connection(bool) parameter to the ControllerPose::DiscoverTrackedDevice() function, which is in charge of choosing the correct eteeController rendermodels position and rotation offsets.

The function should only choose the prior 3DPT offsets (adapted for an eteeeAdaptor (Smart) pose), if the tracker is a VIVE/Tundra and the adaptor_connection(bool) is true. However, this hasn't been tested yet.
TODO: Test the implementation.