feat(3dpt): Apply different offsets to third-party trackers when using Smart of Basic eteeAdaptors
Different position and rotation offsets are applied to the controller rendermodels (in relation to the 3rd-party trackers) depending on whether we use a Smart or Basic eteeAdaptor with the 3rd-party trackers.

Note: Smart adaptor will produce a adaptorConnection = True when connected, while the Basic adaptor is just a 3D print, so no change will happen to the adaptorConnection value when used. This is how we differentiate both.