feat(diff-poses): Allow for two different 'Default' etee bindings
We'll start releasing 2 sets of bindings per game. This is necessary for the wrist offset fixes to work with both eteeTrackers and eteeAdaptor Smart setups.

Variations released for VRChat:
- [eteeTracker & AVR Tundra Tracker] --> VRChat pose
- [AVR VIVE Tracker] --> Raw pose

What the user will need to do: In the Binding Menu, they will just have to click 'Activate' on either one of the official bindings. This depends on which tracker they are using.