fix(error-log): Fix red error printouts in Web Console
Error printout: "[Driver Input] UpdateScalarComponent failed. Invalid handle ulComponent: 0"
Solution: Disable any instance calling ComponentIndex::TRACKPAD_VALUE, as it is not required as an input. This does not affect any functionalities, as the input was already removed from the etee device. The error occured due to us trying to update a scalar component that was not created.